+++ client/src/components/Product/Product.jsx
@@ -4,6 +4,7 @@ import style from './Product.module.css'
 
 function Product({ item }) {
   const navigate = useNavigate()
+  console.log(item);
   return (
     <div onClick={()=> navigate(`/cardProduct/${item.id}`)} className={style.card}>
       <div className={style.blocImg}>
