+++ client/src/redux/reduxThunk/adminThunk.js
@@ -1,5 +1,4 @@
 import {
-  addAdminAC,
   addPhotoAC,
   initActivityAC,
   initColorAC,
@@ -9,17 +8,7 @@ import {
   editPhotoAC,
 } from "../actionCreators/adminAC";
 
-export const addProductFetch = (payload) => {
-  return (dispatch) => {
-    fetch("/createProduct", {
-      method: "POST",
-      headers: { "Content-Type": "Application/json" },
-      body: JSON.stringify(payload),
-    })
-      .then((res) => res.json())
-      .then((data) => dispatch(addAdminAC(data)));
-  };
-};
+
 
 export const searchProductFetch = (payload) => {
   return (dispatch) => {
