+++ client/src/redux/reduxThunk/productsThunk.js
@@ -1,3 +1,4 @@
+import { addAdminAC } from "../actionCreators/adminAC";
 import { initProductsAC } from "../actionCreators/productsAc";
 
 export const initProductsFetch = () => {
@@ -5,5 +6,16 @@ export const initProductsFetch = () => {
     fetch("/allcollection")
       .then((res) => res.json())
       .then((data) => dispatch(initProductsAC(data)));
+  };
+};
+export const addProductFetch = (payload) => {
+  return (dispatch) => {
+    fetch("/createProduct", {
+      method: "POST",
+      headers: { "Content-Type": "Application/json" },
+      body: JSON.stringify(payload),
+    })
+      .then((res) => res.json())
+      .then((data) => dispatch(addAdminAC(data)));
   };
 };
